Go back to Minitest-compatible includes assertion, move compatiblity mixin out of Assertions
