+++ lib/tldr.rb
@@ -10,6 +10,7 @@ require_relative "tldr/hooks"
 require_relative "tldr/parallel_controls"
 require_relative "tldr/path_util"
 require_relative "tldr/planner"
+require_relative "tldr/minitest_compatibility"
 require_relative "tldr/reporters"
 require_relative "tldr/ruby_util"
 require_relative "tldr/runner"
